feat: Ensure only one view state can be active at a time
Prevent having a window which is expanded and minimized at the same time
by removing the class for the previous state upon setting a new state.

This would become annoying to maintain with a large number of states,
but since there are only two here (and the project is for learning
purposes and near completion), I will avoid refactoring it for now.